Rearranging
Moved duplicate functions to Util files to allow for easier reuse.

In PlayerUtil, findClosest Target has been updated to use suared distance.

As far as I can tell, nothing has broken.
