+++ src/main/java/com/gamesense/api/util/player/PlayerUtil.java
@@ -1,0 +1,117 @@
+package com.gamesense.api.util.player;
+
+import com.gamesense.api.util.player.friends.Friends;
+import com.gamesense.api.util.world.EntityUtil;
+import net.minecraft.client.Minecraft;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.util.math.AxisAlignedBB;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Vec3d;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class PlayerUtil {
+    private static final Minecraft mc = Minecraft.getMinecraft();
+
+    public static BlockPos getPlayerPos() {
+        return new BlockPos(Math.floor(mc.player.posX), Math.floor(mc.player.posY), Math.floor(mc.player.posY));
+    }
+
+    // Find closest target
+    // 0b00101010: replaced getDistance with getDistanceSq as speeds up calculation
+    public static EntityPlayer findClosestTarget(double rangeMax, EntityPlayer aimTarget){
+        rangeMax *= rangeMax;
+        List<EntityPlayer> playerList = mc.world.playerEntities;
+
+        EntityPlayer closestTarget = null;
+
+        for (EntityPlayer entityPlayer : playerList){
+
+            if (EntityUtil.basicChecksEntity(entityPlayer))
+                continue;
+
+            if (aimTarget == null && mc.player.getDistanceSq(entityPlayer) <= rangeMax){
+                closestTarget = entityPlayer;
+                continue;
+            }
+            if (aimTarget != null && mc.player.getDistanceSq(entityPlayer) <= rangeMax && mc.player.getDistanceSq(entityPlayer) < mc.player.getDistanceSq(aimTarget)){
+                closestTarget = entityPlayer;
+            }
+        }
+        return closestTarget;
+    }
+
+    // 0b00101010: replaced getDistance with getDistanceSq as speeds up calculation
+    public static EntityPlayer findClosestTarget() {
+        List<EntityPlayer> playerList = mc.world.playerEntities;
+
+        EntityPlayer closestTarget = null;
+
+        for (EntityPlayer entityPlayer : playerList) {
+            if (entityPlayer == mc.player) {
+                continue;
+            }
+            if (Friends.isFriend(entityPlayer.getName())) {
+                continue;
+            }
+            if (entityPlayer.isDead) {
+                continue;
+            }
+            if (closestTarget == null) {
+                closestTarget = entityPlayer;
+                continue;
+            }
+            if (mc.player.getDistanceSq(entityPlayer) < mc.player.getDistanceSq(closestTarget)) {
+                closestTarget = entityPlayer;
+            }
+        }
+
+        return closestTarget;
+    }
+
+    // Find player you are looking
+    public static EntityPlayer findLookingPlayer(double rangeMax) {
+        // Get player
+        ArrayList<EntityPlayer> listPlayer = new ArrayList<>();
+        // Only who is in a distance of enemyRange
+        for(EntityPlayer playerSin : mc.world.playerEntities) {
+            if (EntityUtil.basicChecksEntity(playerSin))
+                continue;
+            if (mc.player.getDistance(playerSin) <= rangeMax) {
+                listPlayer.add(playerSin);
+            }
+        }
+
+        EntityPlayer target = null;
+        // Get coordinate eyes + rotation
+        Vec3d positionEyes = mc.player.getPositionEyes(mc.getRenderPartialTicks());
+        Vec3d rotationEyes = mc.player.getLook(mc.getRenderPartialTicks());
+        // Precision
+        int precision = 2;
+        // Iterate for every blocks
+        for(int i = 0; i < (int) rangeMax; i++) {
+            // Iterate for the precision
+            for(int j = precision; j > 0 ; j--) {
+                // Iterate for all players
+                for(EntityPlayer targetTemp : listPlayer) {
+                    // Get box of the player
+                    AxisAlignedBB playerBox = targetTemp.getEntityBoundingBox();
+                    // Get coordinate of the vec3d
+                    double xArray = positionEyes.x + (rotationEyes.x * i) + rotationEyes.x/j;
+                    double yArray = positionEyes.y + (rotationEyes.y * i) + rotationEyes.y/j;
+                    double zArray = positionEyes.z + (rotationEyes.z * i) + rotationEyes.z/j;
+                    // If it's inside
+                    if (   playerBox.maxY >= yArray && playerBox.minY <= yArray
+                            && playerBox.maxX >= xArray && playerBox.minX <= xArray
+                            && playerBox.maxZ >= zArray && playerBox.minZ <= zArray) {
+                        // Get target
+                        target = targetTemp;
+                    }
+                }
+            }
+        }
+
+        return target;
+    }
+}
