+++ src/main/java/com/gamesense/client/module/modules/hud/CombatInfo.java
@@ -8,6 +8,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import com.gamesense.api.setting.Setting;
+import com.gamesense.api.util.combat.CrystalUtil;
 import com.gamesense.api.util.player.friends.Friends;
 import com.gamesense.api.util.render.GSColor;
 import com.gamesense.client.module.HUDModule;
@@ -63,26 +64,25 @@ public class CombatInfo extends HUDModule {
                 .orElse(null);
     	list.renderLby=false;
     	List<EntityPlayer> entities = new ArrayList<EntityPlayer>(mc.world.playerEntities.stream().filter(entityPlayer -> !Friends.isFriend(entityPlayer.getName())).collect(Collectors.toList()));
-    	AutoCrystalGS a = (AutoCrystalGS) ModuleManager.getModuleByName("AutocrystalGS");
     	for (EntityPlayer e: entities) {
             int i = 0;
             for (BlockPos add: surroundOffset) {
                 ++i;
                 BlockPos o = new BlockPos(e.getPositionVector().x, e.getPositionVector().y, e.getPositionVector().z).add(add.getX(), add.getY(), add.getZ());
                 if (mc.world.getBlockState(o).getBlock() == Blocks.OBSIDIAN) {
-                    if (i == 1 && a.canPlaceCrystal(o.north(1).down())) {
+                    if (i == 1 && CrystalUtil.canPlaceCrystal(o.north(1).down(), AutoCrystalGS.endCrystalMode.getValue())) {
                         list.lby=true;
                         list.renderLby=true;
                     }
-                    else if (i == 2 && a.canPlaceCrystal(o.east(1).down())) {
+                    else if (i == 2 && CrystalUtil.canPlaceCrystal(o.east(1).down(), AutoCrystalGS.endCrystalMode.getValue())) {
                     	list.lby=true;
                         list.renderLby=true;
                     }
-                    else if (i == 3 && a.canPlaceCrystal(o.south(1).down())) {
+                    else if (i == 3 && CrystalUtil.canPlaceCrystal(o.south(1).down(), AutoCrystalGS.endCrystalMode.getValue())) {
                     	list.lby=true;
                         list.renderLby=true;
                     }
-                    else if (i == 4 && a.canPlaceCrystal(o.west(1).down())) {
+                    else if (i == 4 && CrystalUtil.canPlaceCrystal(o.west(1).down(), AutoCrystalGS.endCrystalMode.getValue())) {
                     	list.lby=true;
                         list.renderLby=true;
                     }
